Per-rate GPU profiles, native i24 DoP output, fp64 pipeline, stability fixes
Major changes:
- Per-rate UI columns: FIR mode (Auto/Boxcar/FIR) and Latency (Auto/16-512)
- Config v15: rate_fir_mode[12], rate_lat[12] per-input settings
- Disable GPU SDM (CPU only) — GPU used for FIR only
- Native i24 DoP output via set_data_fixedpoint_ex (was float set_data)
- DoP marker phase continuity across chunks
- FP64 internal pipeline: fir_buf, boxcar, gain all double precision
- SDM/PreCorr input changed to const double* (internals unchanged)
- GPU FIR lowpass kernel (CUDA) for same-rate pre-SDM filtering
- Anti-core-parking: SetThreadExecutionState during playback
- Optimal NTFs from sweep: DSD64→CLANS6, DSD128→SDM6, DSD256→CLANS6
- Optimal latencies: DSD128=128 (+24dB), DSD512=32 (+13dB)
- nc=2 for all same-rate: eliminates candidate collapse accumulation
- DSD512 same-rate uses boxcar (FIR lowpass cutoff too narrow)
- Test logging to foo_dsd_trellis_test.log
- Pipeline SINAD test and dsd_reencode comparison tool

+++ include/precorr.h
@@ -52,7 +52,7 @@ int precorr_context_init(precorr_context_t *ctx, const ntf_filter_t *filter);
  * Input: FIR-upsampled float samples. Output: ±1.0f DSD bits.
  * Returns the number of output samples written (== count, no latency). */
 size_t precorr_process_block(precorr_context_t *ctx,
-                             const float *in, float *out, size_t count);
+                             const double *in, float *out, size_t count);
 
 /* Drain remaining samples (PreCorr has no latency, always returns 0). */
 size_t precorr_drain(precorr_context_t *ctx, float *out, size_t max_out);
